% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proximities.R
\name{as.rf_proximities}
\alias{as.rf_proximities}
\title{as.rf_proximities is a constructor for the rf_proximities class}
\usage{
as.rf_proximities(x)
}
\arguments{
\item{x}{a matrix to be assigned the class rf_proximities}
}
\value{
A matrix of class rf_proximities
}
\description{
as.rf_proximities is a constructor for the rf_proximities class
}
