% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proximities.R
\name{get_proximities}
\alias{get_proximities}
\title{This function generates random forest proximities of a variety of types.}
\usage{
get_proximities(
  x,
  y = NULL,
  rf = NULL,
  x_test = NULL,
  type = "rfgap",
  nk = 10,
  ...
)
}
\arguments{
\item{x}{A dataframe or matrix of data. Rows (n) are observations,
columns (d) are variables.}

\item{y}{The labels corresponding to x. should be of type factor.}

\item{rf}{A forest object of s3 class 'ranger'. Note: rf requires
write.forest = TRUE and keep.inbag = TRUE.}

\item{x_test}{An (optional) test dataset (dataframe or matrix)}

\item{type}{The type of proximities to be obtained.
Options are rfgap (default), original, oob, pbk, or rfproxih.}

\item{nk}{Number of terminal node neighbors to consider when calling type =
rfproxih. (default 10).}

\item{...}{Additional argument options for ranger(). (If rf is not supplied).}
}
\value{
A matrix of proximity values.
}
\description{
This function generates random forest proximities of a variety of types.
}
\examples{

x <- iris[, 1:4]
y <- iris[, 5]

prox <- get_proximities(x, y)
}
