% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes_distance.R
\name{nodes_distance}
\alias{nodes_distance}
\title{Pranger}
\usage{
nodes_distance(parent_nodes, node1, node2)
}
\arguments{
\item{parent_nodes}{[\code{data.table}] \code{data.table} of parent nodes}

\item{node1}{[\code{integer}] A node}

\item{node2}{[\code{integer}] A node}
}
\value{
[\code{integer}] distance between node1 and node2
}
\description{
Compute distance between two terminal nodes
}
\author{
Cesaire J. K. Fouodo
}
