% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_mds.R
\name{plot.rf_mds}
\alias{plot.rf_mds}
\title{This is is a generic plot function for an rf_mds object}
\usage{
\method{plot}{rf_mds}(x, y, ...)
}
\arguments{
\item{x}{an rf_mds object}

\item{y}{class labels associated with x.}

\item{...}{Additional arguments for generic plot function}
}
\description{
This is is a generic plot function for an rf_mds object
}
