% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_outliers.R
\name{plot.rf_outlier}
\alias{plot.rf_outlier}
\title{This is is a generic plot function for an rf_outlier object}
\usage{
\method{plot}{rf_outlier}(outlier, x, y, base_size = 2, scale_size = 3, ...)
}
\arguments{
\item{outlier}{an rf_outlier object}

\item{x}{dataframe associated with the outlier scores}

\item{y}{class labels associated with x.}

\item{base_size}{starting point size}

\item{scale_size}{how much is the outlier score affecting point size}

\item{...}{Additional arguments for mds and generic plotting}
}
\description{
This is is a generic plot function for an rf_outlier object
}
