% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rf_proximities}
\alias{predict.rf_proximities}
\title{A function to make predictions using random forest proximities
 (rf_proximities class)}
\usage{
\method{predict}{rf_proximities}(object, y, ...)
}
\arguments{
\item{object}{An rf_proximity class matrix}

\item{y}{A vector of class labels or a continuous responses}

\item{...}{Additional arguments for running ranger::ranger}
}
\value{
A list which contains the predictions and prediction error rate
}
\description{
A function to make predictions using random forest proximities
 (rf_proximities class)
}
\examples{
x <- iris[, 1:4]
y <- iris[, 5]
prox <- get_proximities(x, y)
preds_list <- predict(prox, y)
}
