% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_impute.R
\name{rf_impute}
\alias{rf_impute}
\title{A function to impute missing values from a dataset.}
\usage{
rf_impute(x, y, type = "rfgap", n_iters = 1, x_true = NULL, ...)
}
\arguments{
\item{x}{A dataframe or matrix of data. Rows (n) are observations,
columns (d) are variables.}

\item{y}{The labels corresponding to x. Should be of type factor.}

\item{type}{The type of proximities to be obtained. Options are
rfgap (default), original, or oob.}

\item{n_iters}{The number of iterations for imputation.  The default is 1.}

\item{x_true}{The x data matrix prior to removing values for experimentation.
This is used to compare imputation methods, not used in "real-life"
applications.}

\item{...}{Additional argument options for ranger(). (If rf is not supplied).}
}
\value{
A dataframe with imputed values if x_true is NULL. Else, a list
  which contains the imputed dataset and error values (MSE between the
  true and imputed data)
}
\description{
A function to impute missing values from a dataset.
}
\examples{
x <- airquality[, 1:4]
y <- airquality[, 5]
x_imp <- rf_impute(x, y)
}
