% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_outliers.R
\name{rf_outliers}
\alias{rf_outliers}
\title{detect_outliers produces a dataset's outlier measure score based on random-forest
  proximities}
\usage{
rf_outliers(x, y, type = "rfgap", rf = NULL, ...)
}
\arguments{
\item{x}{A data matrix (rows are observations, columns variables) or a
pre-computed proximity matrix of class rf_proximities.}

\item{y}{A vector of class labels or a continuous responses}

\item{type}{The type of random forest proximity to be used.  Options are
'rfgap', 'oob', or 'original'.}

\item{rf}{(Optional) a pretrained random forest of class ranger.}

\item{...}{Additional options for running ranger::ranger()}
}
\value{
A vector of outlier measures of length length(y)
}
\description{
detect_outliers produces a dataset's outlier measure score based on random-forest
  proximities
}
\examples{
x <- iris[, 1:4]
y <- iris[, 5]
outlier_measures <- rf_outliers(x, y)
}
