% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_parents.R
\name{search_parents}
\alias{search_parents}
\title{Pranger}
\usage{
search_parents(ranger_tree)
}
\arguments{
\item{ranger_tree}{[\code{ranger.tree}] A ranger tree\cr}
}
\value{
[\code{data.table}]
  A \code{data.table} containing parent nodes of each terminal node
}
\description{
Search parent nodes of each terminal node from a tree
}
\author{
Cesaire J. K. Fouodo
}
