% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance.R
\name{tree_distance}
\alias{tree_distance}
\title{Pranger}
\usage{
tree_distance(ranger_tree)
}
\arguments{
\item{ranger_tree}{[\code{ranger.tree}] A ranger tree}
}
\value{
[\code{matrix}] A matrix of dissimilarities
}
\description{
Compute distances between all terminal nodes for a given tree
}
\author{
Cesaire J. K. Fouodo
}
